backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
psycopg[binary]>=3.2.10
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.1
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
slowapi>=0.1.9
alembic>=1.13.0
httpx>=0.27.0
python-multipart>=0.0.9
